style: enhance color contrast for medication cards (accessibility)
